requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.14
pillow==10.0.1
numpy>=1.26.0
